add info on how to create a new variant sign in gate
