+++ static/src/javascripts/projects/common/modules/identity/sign-in-gate/README.md
@@ -1,0 +1,16 @@
+# Sign In Gate Test
+## Adding Variants
+
+1. Add the new variant to the sign in gate ab test definition in the `variants` array: `static/src/javascripts/projects/common/modules/experiments/tests/sign-in-gate.js`
+2. Clone the `example.js` file in the `./variants` folder, have the name of the file reflect the variant name, e.g. `control.js`.
+3. Modify this file to reflect the new variant information.
+    - First define the variant name, by changing `example` in `const name = 'example'` to the new variant name
+    - Next change the `htmlTemplate` method to return the new template for the variant
+    - Then modify the `canShow` method, this method determines if the test can be shown on that page view for that variant
+    - After modify the `show` method to run anything thats needed to get the gate to show, e.g. setting the template, adding click handlers etc.
+    - Finally export a `SignInGateVariant` type, which is an object that exports `name`, `canShow`, and `show`.
+4. Import this object in `./variants/index.js`, and add it to the export array. From this point on the variant will be able to be used.
+
+## Styling the game
+
+CSS for the gate lives at `static/src/stylesheets/module/identity/_sign-in-gate.scss`, modify or add anything here.
